// Frontend/src/models/CreateSalesforceAccountRequest.ts
export interface CreateSalesforceAccountRequest {
    userId: string;
    accountName: string;
    accountPhoneNumber: string;
    accountWebsite: string;
    contactFirstName: string;
    contactLastName: string;
    contactEmail: string;
    contactPhoneNumber: string;
    contactTitle: string;
}